Updated Ant Design. This caused a lot of bugs so this commit includes a lot of refactoring/bug fixing.
Next to that it is now also possible to add up to four images to a product. These images are saved to an external third party image server (Cloudinary). URLs to these images are being saved in the database. The images are in a carousel on the home page.

I also changed the shown text to be Dutch and made some styling changes.